feat(ai): add basic scrollycoding components from codehike
