+++ apps/ai-hero/src/components/codehike/code.tsx
@@ -1,11 +1,17 @@
-import { AnnotationHandler, InnerToken } from 'codehike/code'
+import React from 'react'
+import { highlight, Pre, RawCode } from 'codehike/code'
 
-import { SmoothPre } from './smooth-pre'
+import { callout } from './callout'
+import { fold } from './fold'
 
-export const tokenTransitions: AnnotationHandler = {
-	name: 'token-transitions',
-	PreWithRef: SmoothPre,
-	Token: (props) => (
-		<InnerToken merge={props} style={{ display: 'inline-block' }} />
-	),
+export async function Code({ codeblock }: { codeblock: RawCode }) {
+	const highlighted = await highlight(codeblock, 'github-dark')
+	return (
+		<Pre
+			code={highlighted}
+			className="bg-background text-xs sm:text-sm"
+			style={{ ...highlighted.style, padding: '1rem', borderRadius: '0.5rem' }}
+			handlers={[callout, fold]}
+		/>
+	)
 }
